refactor: extract MicrophoneLevelTypes.swift from MenuBarView.swift
Extracts three pure-move types from MenuBarView.swift into a new file:
- MicrophoneInputLevelState (equatable enum with label + accessibilityValue)
- MicrophoneLevelCalculator (normalizedRMSLevel(for:) on AVAudioPCMBuffer)
- MicrophoneInputLevelTapFactory (makeTap wrapper)

Adds `import AVFoundation` (needed for AVAudioPCMBuffer / AVAudioNodeTapBlock).

Byte-identical extraction of lines 805-877 from MenuBarView.swift.
Verified with SHA-256 of the extracted range.

Reduces MenuBarView.swift from 877 to 803 lines; new file is 75 lines.
Tests: 667 (unchanged).
